{dist,sample} => {dist,sample}uv: rename packages
