+++ distuv/interfaces.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package dist
+package distuv
 
 type LogProber interface {
 	LogProb(float64) float64
