+++ sampleuv/weighted.go
@@ -2,7 +2,7 @@
 // Use of this code is governed by a BSD-style
 // license that can be found in the LICENSE file
 
-package sample
+package sampleuv
 
 import (
 	"math/rand"
